docs: document all interfaces and classes
