--- packages/core/src/cache-container/cache-container-types.ts
@@ -1,28 +1,0 @@
-// eslint-disable-next-line @typescript-eslint/no-explicit-any
-export type CachedItem<T = any> = {
-  content: T;
-  meta: {
-    createdAt: number;
-    ttl: number | null;
-    isLazy: boolean;
-  };
-};
-
-export type CachingOptions = {
-  /** (Default: 60) Number of seconds to expire the cachte item */
-  ttl: number;
-  /** (Default: true) If true, expired cache entries will be deleted on touch and returned anyway. If false, entries will be deleted after the given ttl. */
-  isLazy: boolean;
-  /** (Default: false) If true, cache entry has no expiration. */
-  isCachedForever: boolean;
-  /** (Default: JSON.stringify combination of className, methodName and call args) */
-  calculateKey: (data: {
-    /** The class name for the method being decorated */
-    className: string;
-    /** The method name being decorated */
-    methodName: string;
-    /** The arguments passed to the method when called */
-    // eslint-disable-next-line @typescript-eslint/no-explicit-any
-    args: any[];
-  }) => string;
-};
