--- packages/core/src/cache-container/index.ts
@@ -1,2 +1,0 @@
-export * from "./cache-container";
-export * from "./cache-container-types";
