--- packages/core/src/decorator/index.ts
@@ -1,1 +1,0 @@
-export * from "./cache-decorator";
