+++ packages/core/src/index.ts
@@ -1,4 +1,4 @@
 export * from "./decorator";
 export * from "./storage";
-export * from "./cache-container";
+export * from "./cacheContainer";
 export * from "./withCache";
