--- packages/core/src/storage/index.ts
@@ -1,1 +1,0 @@
-export * from "./storage-types";
