+++ packages/storage-ioredis/src/ioredis-storage.test.ts
@@ -1,7 +1,7 @@
 import type { default as IORedis } from "ioredis";
 import { IoRedisStorage } from ".";
-import { storageTestFactory } from "../../core/src/storage/storageTestFactory";
-import { decoratorTestFactory } from "../../core/src/decorator/decoratorTestFactory";
+import { storageTestFactory } from "../../core/src/utils/storageTestFactory";
+import { decoratorTestFactory } from "../../core/src/utils/decoratorTestFactory";
 // eslint-disable-next-line @typescript-eslint/no-var-requires
 const IoRedisMock: typeof IORedis = require("ioredis-mock");
