+++ packages/storage-memory/src/memory-storage.test.ts
@@ -1,6 +1,6 @@
 import { MemoryStorage } from ".";
-import { storageTestFactory } from "../../core/src/storage/storageTestFactory";
-import { decoratorTestFactory } from "../../core/src/decorator/decoratorTestFactory";
+import { storageTestFactory } from "../../core/src/utils/storageTestFactory";
+import { decoratorTestFactory } from "../../core/src/utils/decoratorTestFactory";
 
 describe("memory-storage", () => {
   const storage = new MemoryStorage();
